requirements.txt

Flask
gspread
google-auth
